Combine unit_test.go to zipkin_test.go
